Move enum map to structural
